move types folder, refactor redux naming to transition to rtk query
