+++ src/components/Dashboard/PackCategory/AddCategoryButton/AddCategoryButton.tsx
@@ -1,24 +1,23 @@
-import { Button, Icon } from "semantic-ui-react";
+import { Button, Icon } from 'semantic-ui-react';
 
-interface AddCategoryButtonProps {
-  onClick: () => void;
-}
+type AddCategoryButtonProps = {
+	onClick: () => void;
+};
 
 const AddCategoryButton = (props: AddCategoryButtonProps) => {
-  return (
-    <div>
-      <Button
-        color="grey"
-        basic
-        size="small"
-        className="add-item-table-button"
-        onClick={props.onClick}
-      >
-        <Icon name="tree" />
-        Add Category
-      </Button>
-    </div>
-  );
+	return (
+		<div>
+			<Button
+				color="grey"
+				basic
+				size="small"
+				className="add-item-table-button"
+				onClick={props.onClick}>
+				<Icon name="tree" />
+				Add Category
+			</Button>
+		</div>
+	);
 };
 
 export default AddCategoryButton;
