+++ src/components/Dashboard/PackCategory/TableButtonCells/QuantityButton.tsx
@@ -1,48 +1,47 @@
-import { Table, Input } from "semantic-ui-react";
-import { useState } from "react";
-import "./TableButtons.css";
+import { Table, Input } from 'semantic-ui-react';
+import { useState } from 'react';
+import './TableButtons.css';
 
-interface ButtonProps {
-  quantity: number;
-  size: number;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  onToggleOff: () => void;
-}
+type ButtonProps = {
+	quantity: number;
+	size: number;
+	onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+	onToggleOff: () => void;
+};
 
 const QuantityButton: React.FC<ButtonProps> = (props: ButtonProps) => {
-  const { quantity, size, onChange, onToggleOff } = props;
+	const { quantity, size, onChange, onToggleOff } = props;
 
-  const [toggleInput, setToggleInput] = useState(false);
-  const toggleToEdit = () => !toggleInput && setToggleInput(true);
-  const toggleToCell = () => {
-    if (toggleInput) {
-      setToggleInput(false);
-      onToggleOff();
-    }
-  };
+	const [toggleInput, setToggleInput] = useState(false);
+	const toggleToEdit = () => !toggleInput && setToggleInput(true);
+	const toggleToCell = () => {
+		if (toggleInput) {
+			setToggleInput(false);
+			onToggleOff();
+		}
+	};
 
-  return (
-    <Table.Cell
-      className="table-button"
-      textAlign="center"
-      colSpan={size}
-      onMouseOver={toggleToEdit}
-      onMouseLeave={toggleToCell}
-      onBlur={toggleToCell}
-      onClick={toggleToEdit}
-    >
-      <Input
-        fluid
-        name="packItemQuantity"
-        value={quantity}
-        type="number"
-        step={1}
-        transparent={!toggleInput}
-        onChange={onChange}
-        style={{ paddingLeft: toggleInput ? "0px" : "12px" }}
-      />
-    </Table.Cell>
-  );
+	return (
+		<Table.Cell
+			className="table-button"
+			textAlign="center"
+			colSpan={size}
+			onMouseOver={toggleToEdit}
+			onMouseLeave={toggleToCell}
+			onBlur={toggleToCell}
+			onClick={toggleToEdit}>
+			<Input
+				fluid
+				name="packItemQuantity"
+				value={quantity}
+				type="number"
+				step={1}
+				transparent={!toggleInput}
+				onChange={onChange}
+				style={{ paddingLeft: toggleInput ? '0px' : '12px' }}
+			/>
+		</Table.Cell>
+	);
 };
 
 export default QuantityButton;
