+++ src/components/Dashboard/PackInfo/PackFormModal/PackFormModal.tsx
@@ -14,17 +14,17 @@ import {
 	Divider,
 } from 'semantic-ui-react';
 import { useState, useEffect } from 'react';
-import { useEditPackMutation } from '../../../../redux/newPacks/newPacksApiSlice';
-import { Pack } from '../../../../redux/packs/packTypes';
+import { useEditPackMutation } from '../../../../redux/pack/packApiSlice';
+import { Pack } from '../../../../types/packTypes';
 import './PackFormModal.css';
 import PackTagProperties from './PackTagProperties/PackTagProperties';
 
-interface PackFormModalProps {
+type PackFormModalProps = {
 	pack: Pack;
 	open: boolean;
 	onClose: () => void;
 	onClickDelete: () => void;
-}
+};
 
 const PackFormModal = (props: PackFormModalProps) => {
 	const [editPack] = useEditPackMutation();
