+++ src/redux/previousPackRedux/packUtils.ts
@@ -1,0 +1,17 @@
+import { PackItem, Category, PackListItem } from '../../types/packTypes';
+
+export const getCategoryIdx = (categories: Category[], categoryId: number) => {
+	return categories.findIndex(
+		(item: Category) => item.packCategoryId === Number(categoryId),
+	);
+};
+
+export const getPackItemIdx = (category: Category, packItemId: number) => {
+	return category.packItems.findIndex(
+		(item: PackItem) => item.packItemId === Number(packItemId),
+	);
+};
+
+export const getPackIdx = (packList: PackListItem[], packId: number) => {
+	return packList.findIndex((pack: PackListItem) => pack.packId === Number(packId));
+};
