+++ src/utils/mockData.ts
@@ -1,111 +1,111 @@
-import { PackItem } from '../redux/packs/packTypes';
+import { PackItem } from '../types/packTypes';
 
 export const mockPackItems: PackItem[] = [
-  {
-    packItemId: 1,
-    packCategoryId: 1,
-    packItemName: 'Backpack',
-    packItemDescription: '40 Liter UL',
-    packItemWeight: 40,
-    packItemUnit: 'oz',
-    packItemQuantity: 1,
-    packItemUrl: '',
-    wornWeight: false,
-    consumable: false,
-    favorite: false,
-  },
-  {
-    packItemId: 2,
-    packCategoryId: 1,
-    packItemName: '',
-    packItemDescription: '',
-    packItemWeight: 1,
-    packItemUnit: 'lb',
-    packItemQuantity: 1,
-    packItemUrl: '',
-    wornWeight: false,
-    consumable: false,
-    favorite: false,
-  },
-  {
-    packItemId: 3,
-    packCategoryId: 1,
-    packItemName: '',
-    packItemDescription: '',
-    packItemWeight: 1.5,
-    packItemUnit: 'kg',
-    packItemQuantity: 1,
-    packItemUrl: '',
-    wornWeight: false,
-    consumable: false,
-    favorite: false,
-  },
-  {
-    packItemId: 4,
-    packCategoryId: 1,
-    packItemName: '',
-    packItemDescription: '',
-    packItemWeight: 500,
-    packItemUnit: 'g',
-    packItemQuantity: 1,
-    packItemUrl: '',
-    wornWeight: false,
-    consumable: false,
-    favorite: false,
-  },
+	{
+		packItemId: 1,
+		packCategoryId: 1,
+		packItemName: 'Backpack',
+		packItemDescription: '40 Liter UL',
+		packItemWeight: 40,
+		packItemUnit: 'oz',
+		packItemQuantity: 1,
+		packItemUrl: '',
+		wornWeight: false,
+		consumable: false,
+		favorite: false,
+	},
+	{
+		packItemId: 2,
+		packCategoryId: 1,
+		packItemName: '',
+		packItemDescription: '',
+		packItemWeight: 1,
+		packItemUnit: 'lb',
+		packItemQuantity: 1,
+		packItemUrl: '',
+		wornWeight: false,
+		consumable: false,
+		favorite: false,
+	},
+	{
+		packItemId: 3,
+		packCategoryId: 1,
+		packItemName: '',
+		packItemDescription: '',
+		packItemWeight: 1.5,
+		packItemUnit: 'kg',
+		packItemQuantity: 1,
+		packItemUrl: '',
+		wornWeight: false,
+		consumable: false,
+		favorite: false,
+	},
+	{
+		packItemId: 4,
+		packCategoryId: 1,
+		packItemName: '',
+		packItemDescription: '',
+		packItemWeight: 500,
+		packItemUnit: 'g',
+		packItemQuantity: 1,
+		packItemUrl: '',
+		wornWeight: false,
+		consumable: false,
+		favorite: false,
+	},
 ];
 
 export const mockPackItemsWithMultipleItems: PackItem[] = [
-  {
-    packItemId: 1,
-    packCategoryId: 1,
-    packItemName: 'Backpack',
-    packItemDescription: '40 Liter UL',
-    packItemWeight: 40,
-    packItemUnit: 'oz',
-    packItemQuantity: 2,
-    packItemUrl: '',
-    wornWeight: false,
-    consumable: false,
-    favorite: false,
-  },
-  {
-    packItemId: 2,
-    packCategoryId: 1,
-    packItemName: '',
-    packItemDescription: '',
-    packItemWeight: 1,
-    packItemUnit: 'lb',
-    packItemQuantity: 3,
-    packItemUrl: '',
-    wornWeight: false,
-    consumable: false,
-    favorite: false,
-  },
-  {
-    packItemId: 3,
-    packCategoryId: 1,
-    packItemName: '',
-    packItemDescription: '',
-    packItemWeight: 1.5,
-    packItemUnit: 'kg',
-    packItemQuantity: 1,
-    packItemUrl: '',
-    wornWeight: false,
-    consumable: false,
-    favorite: false,
-  },
-  {
-    packItemId: 4,
-    packCategoryId: 1,
-    packItemName: '',
-    packItemDescription: '',
-    packItemWeight: 500,
-    packItemUnit: 'g',
-    packItemQuantity: 1,
-    packItemUrl: '',
-    wornWeight: false,
-    consumable: false,
-    favorite: false,
-  },
+	{
+		packItemId: 1,
+		packCategoryId: 1,
+		packItemName: 'Backpack',
+		packItemDescription: '40 Liter UL',
+		packItemWeight: 40,
+		packItemUnit: 'oz',
+		packItemQuantity: 2,
+		packItemUrl: '',
+		wornWeight: false,
+		consumable: false,
+		favorite: false,
+	},
+	{
+		packItemId: 2,
+		packCategoryId: 1,
+		packItemName: '',
+		packItemDescription: '',
+		packItemWeight: 1,
+		packItemUnit: 'lb',
+		packItemQuantity: 3,
+		packItemUrl: '',
+		wornWeight: false,
+		consumable: false,
+		favorite: false,
+	},
+	{
+		packItemId: 3,
+		packCategoryId: 1,
+		packItemName: '',
+		packItemDescription: '',
+		packItemWeight: 1.5,
+		packItemUnit: 'kg',
+		packItemQuantity: 1,
+		packItemUrl: '',
+		wornWeight: false,
+		consumable: false,
+		favorite: false,
+	},
+	{
+		packItemId: 4,
+		packCategoryId: 1,
+		packItemName: '',
+		packItemDescription: '',
+		packItemWeight: 500,
+		packItemUnit: 'g',
+		packItemQuantity: 1,
+		packItemUrl: '',
+		wornWeight: false,
+		consumable: false,
+		favorite: false,
+	},
 ];
