+++ src/utils/weightConverter.ts
@@ -1,90 +1,90 @@
-import { PackItem } from '../redux/packs/packTypes';
+import { PackItem } from '../types/packTypes';
 
 export const weightConverter = (itemList: PackItem[], outputUnit: string) => {
-  if (itemList[0]) {
-    const newItem = itemList
-      .map(
-        (item: {
-          packItemWeight: number;
-          packItemUnit: string;
-          packItemQuantity: number;
-        }) => {
-          const { packItemWeight, packItemUnit, packItemQuantity } = item;
-          // Ensure we always return number type for reducer
-          const itemWeight = Number(packItemWeight) * packItemQuantity;
+	if (itemList[0]) {
+		const newItem = itemList
+			.map(
+				(item: {
+					packItemWeight: number;
+					packItemUnit: string;
+					packItemQuantity: number;
+				}) => {
+					const { packItemWeight, packItemUnit, packItemQuantity } = item;
+					// Ensure we always return number type for reducer
+					const itemWeight = Number(packItemWeight) * packItemQuantity;
 
-          if (packItemUnit === outputUnit || itemWeight === 0) {
-            return itemWeight;
-          }
-          if (outputUnit === 'oz') {
-            return convertToOunces(itemWeight, packItemUnit);
-          } else if (outputUnit === 'lb') {
-            return convertToPounds(itemWeight, packItemUnit);
-          } else if (outputUnit === 'kg') {
-            return convertToKilograms(itemWeight, packItemUnit);
-          } else if (outputUnit === 'g') {
-            return convertToGrams(itemWeight, packItemUnit);
-          } else return itemWeight;
-        },
-      )
-      .reduce((weight: number, sum: number = 0) => (sum += weight))
-      .toFixed(2);
-    return Number(newItem);
-  } else return 0;
+					if (packItemUnit === outputUnit || itemWeight === 0) {
+						return itemWeight;
+					}
+					if (outputUnit === 'oz') {
+						return convertToOunces(itemWeight, packItemUnit);
+					} else if (outputUnit === 'lb') {
+						return convertToPounds(itemWeight, packItemUnit);
+					} else if (outputUnit === 'kg') {
+						return convertToKilograms(itemWeight, packItemUnit);
+					} else if (outputUnit === 'g') {
+						return convertToGrams(itemWeight, packItemUnit);
+					} else return itemWeight;
+				},
+			)
+			.reduce((weight: number, sum: number = 0) => (sum += weight))
+			.toFixed(2);
+		return Number(newItem);
+	} else return 0;
 };
 
 function convertToOunces(weight: number, unit: string) {
-  switch (unit) {
-    case 'lb':
-      return weight * 16;
-    case 'kg':
-      return weight * 0.0283495;
-    case 'g':
-      return weight * 28.3495;
-    default:
-      return weight;
-  }
+	switch (unit) {
+		case 'lb':
+			return weight * 16;
+		case 'kg':
+			return weight * 0.0283495;
+		case 'g':
+			return weight * 28.3495;
+		default:
+			return weight;
+	}
 }
 
 function convertToPounds(weight: number, unit: string) {
-  switch (unit) {
-    case 'oz':
-      return weight / 16;
-    case 'kg':
-      return weight * 2.20462;
-    case 'g':
-      return weight * 0.00220462;
-    default:
-      return weight;
-  }
+	switch (unit) {
+		case 'oz':
+			return weight / 16;
+		case 'kg':
+			return weight * 2.20462;
+		case 'g':
+			return weight * 0.00220462;
+		default:
+			return weight;
+	}
 }
 
 function convertToKilograms(weight: number, unit: string) {
-  switch (unit) {
-    case 'oz':
-      return weight * 0.0283495;
-    case 'lb':
-      return weight * 0.453592;
-    case 'g':
-      return weight / 1000;
-    default:
-      return weight;
-  }
+	switch (unit) {
+		case 'oz':
+			return weight * 0.0283495;
+		case 'lb':
+			return weight * 0.453592;
+		case 'g':
+			return weight / 1000;
+		default:
+			return weight;
+	}
 }
 
 function convertToGrams(weight: number, unit: string) {
-  switch (unit) {
-    case 'oz':
-      return weight * 28.3495;
-    case 'lb':
-      return weight * 453.592;
-    case 'kg':
-      return weight * 1000;
-    default:
-      return weight;
-  }
+	switch (unit) {
+		case 'oz':
+			return weight * 28.3495;
+		case 'lb':
+			return weight * 453.592;
+		case 'kg':
+			return weight * 1000;
+		default:
+			return weight;
+	}
 }
 
 export const quantityConverter = (items: PackItem[]) => {
-  return items.reduce((acc, item) => (acc += item.packItemQuantity), 0);
+	return items.reduce((acc, item) => (acc += item.packItemQuantity), 0);
 };
